pyproject.toml
[project]
name = "toolhoadinh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "dotenv>=0.9.9",
    "pandas>=2.3.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
